解决CAN接收扩展帧后发送全消失的问题，优化了CAN驱动代码，各功能完好

+++ BMS/driver/BMS/PrechargeM.c
@@ -51,7 +51,7 @@ int PrechargeM_IsFail()             //超时判断函数
 int PrechargeM_IsFinish()           //预充完成判断函数
 {
     int rebuf;
-    if (Hv_Get(Pre_CfgState.pre_voltage ,HV_VOLTAGE) * 100 >=
+    if (Hv_Get((uint8)Pre_CfgState.pre_voltage ,HV_VOLTAGE) * 100 >=
         Hv_Get(0 ,HV_VOLTAGE) * (Pre_CfgState.percent))   //判断V1电压是否大于BAT电压的95%
     {
         rebuf = 1;
